fixed error with wrong procrustes distance results and scaling of model (size was too large)
