+++ GUI/src/cz/fidentis/gui/comparison_batch/Bundle.properties
@@ -41,7 +41,6 @@ ExportPanel.jLabel3.text=Export points for:
 BatchComparisonResults.jButton8.text=Export auxiliary results...
 BatchComparisonResults.jButton9.text=Export average face...
 BatchComparisonConfiguration.processComparisonButton.text=Process Comparison
-BatchComparisonConfiguration.jLabel8.text=Enable scaling:
 BatchComparisonConfiguration.jLabel3.text=Comparison:
 BatchComparisonConfiguration.jLabel2.text=Comparison
 BatchRegistrationConfiguration.jLabel8.text=Average meshes:
@@ -120,7 +119,6 @@ BatchRegistrationConfiguration.fpPointInfoCheckBox.text=show point info
 BatchRegistrationConfiguration.fpSizeSlider.toolTipText=
 BatchRegistrationConfiguration.icpScaleCheckBox.text=
 BatchRegistrationConfiguration.continueComparisonCheckBox.text=Continue directly with comparison
-BatchComparisonConfiguration.fpScaleCheckBox.text=
 BatchComparisonResults.metricComboBox.toolTipText=
 BatchComparisonResults.xyRadioButton.text=XY
 BatchComparisonResults.yzRadioButton.text=YZ
