+++ GUI/src/cz/fidentis/gui/comparison_one_to_many/Bundle.properties
@@ -40,7 +40,6 @@ ExportPanel.jLabel2.text=Folder:
 ExportPanel.jTextField2.text=
 OneToManyComparisonConfiguration.jLabel2.text=Comparison
 OneToManyComparisonConfiguration.jLabel12.text=Threshold:
-OneToManyComparisonConfiguration.jLabel8.text=Enable scaling:
 OneToManyComparisonConfiguration.processComparisonButton.text=Process Comparison
 OneToManyComparisonConfiguration.jLabel3.text=Comparison:
 # To change this template, choose Tools | Templates
@@ -116,10 +115,6 @@ OneToManyRegistrationConfiguration.removeFPButton.text=Remove
 OneToManyRegistrationConfiguration.editFPButton.text=Edit
 OneToManyRegistrationConfiguration.validateFPButton.toolTipText=Check if points lie within normal variation.
 OneToManyRegistrationConfiguration.validateFPButton.text=Validate 
-# To change this license header, choose License Headers in Project Properties.
-# To change this template file, choose Tools | Templates
-# and open the template in the editor.
-OneToManyComparisonConfiguration.fpScalingCheckBox.text=
 OneToManyComparisonResults.metricComboBox.toolTipText=
 OneToManyComparisonResults.xyPlaneRadioButton.text=XY
 OneToManyComparisonResults.yzPlaneRadioButton.text=YZ
