+++ GUI/src/cz/fidentis/gui/comparison_one_to_many/OneToManyComparisonConfiguration.java
@@ -78,8 +78,6 @@ public class OneToManyComparisonConfiguration extends javax.swing.JPanel {
         filler1 = new javax.swing.Box.Filler(new java.awt.Dimension(0, 0), new java.awt.Dimension(0, 0), new java.awt.Dimension(32767, 32767));
         jPanel5 = new javax.swing.JPanel();
         procrustesPanel = new javax.swing.JPanel();
-        fpScalingCheckBox = new javax.swing.JCheckBox();
-        jLabel8 = new javax.swing.JLabel();
         fpThresholdSlider = new javax.swing.JSlider();
         jLabel12 = new javax.swing.JLabel();
         processComparisonButton = new javax.swing.JButton();
@@ -106,15 +104,6 @@ public class OneToManyComparisonConfiguration extends javax.swing.JPanel {
 
         procrustesPanel.setVisible(false);
 
-        org.openide.awt.Mnemonics.setLocalizedText(fpScalingCheckBox, org.openide.util.NbBundle.getMessage(OneToManyComparisonConfiguration.class, "OneToManyComparisonConfiguration.fpScalingCheckBox.text")); // NOI18N
-        fpScalingCheckBox.addChangeListener(new javax.swing.event.ChangeListener() {
-            public void stateChanged(javax.swing.event.ChangeEvent evt) {
-                fpScalingCheckBoxStateChanged(evt);
-            }
-        });
-
-        org.openide.awt.Mnemonics.setLocalizedText(jLabel8, org.openide.util.NbBundle.getMessage(OneToManyComparisonConfiguration.class, "OneToManyComparisonConfiguration.jLabel8.text")); // NOI18N
-
         fpThresholdSlider.setMajorTickSpacing(20);
         fpThresholdSlider.setMinorTickSpacing(5);
         fpThresholdSlider.setPaintLabels(true);
@@ -133,24 +122,15 @@ public class OneToManyComparisonConfiguration extends javax.swing.JPanel {
         procrustesPanelLayout.setHorizontalGroup(
             procrustesPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(procrustesPanelLayout.createSequentialGroup()
-                .addGroup(procrustesPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(procrustesPanelLayout.createSequentialGroup()
-                        .addComponent(jLabel12, javax.swing.GroupLayout.PREFERRED_SIZE, 61, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addGap(18, 18, 18)
-                        .addComponent(fpThresholdSlider, javax.swing.GroupLayout.PREFERRED_SIZE, 187, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addGroup(procrustesPanelLayout.createSequentialGroup()
-                        .addComponent(jLabel8)
-                        .addGap(18, 18, 18)
-                        .addComponent(fpScalingCheckBox)))
+                .addComponent(jLabel12, javax.swing.GroupLayout.PREFERRED_SIZE, 61, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(18, 18, 18)
+                .addComponent(fpThresholdSlider, javax.swing.GroupLayout.PREFERRED_SIZE, 187, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
         procrustesPanelLayout.setVerticalGroup(
             procrustesPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(procrustesPanelLayout.createSequentialGroup()
-                .addGroup(procrustesPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                    .addComponent(jLabel8, javax.swing.GroupLayout.PREFERRED_SIZE, 24, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(fpScalingCheckBox))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addContainerGap()
                 .addGroup(procrustesPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                     .addComponent(jLabel12, javax.swing.GroupLayout.PREFERRED_SIZE, 24, javax.swing.GroupLayout.PREFERRED_SIZE)
                     .addComponent(fpThresholdSlider, javax.swing.GroupLayout.PREFERRED_SIZE, 46, javax.swing.GroupLayout.PREFERRED_SIZE)))
@@ -294,10 +274,6 @@ public class OneToManyComparisonConfiguration extends javax.swing.JPanel {
     private void fpThresholdSliderStateChanged(javax.swing.event.ChangeEvent evt) {//GEN-FIRST:event_fpThresholdSliderStateChanged
         getContext().setFpTreshold(fpThresholdSlider.getValue());
     }//GEN-LAST:event_fpThresholdSliderStateChanged
-
-    private void fpScalingCheckBoxStateChanged(javax.swing.event.ChangeEvent evt) {//GEN-FIRST:event_fpScalingCheckBoxStateChanged
-        getContext().setFpScaling(fpScalingCheckBox.isSelected());
-    }//GEN-LAST:event_fpScalingCheckBoxStateChanged
 
     public void computeComparison(final ProjectTopComponent tc){
         final OneToManyComparison c = getContext();
@@ -564,7 +540,6 @@ public class OneToManyComparisonConfiguration extends javax.swing.JPanel {
                 
         compMethodComboBox.setSelectedItem(c.getComparisonMethod());
         createAvgCheckBox.setSelected(c.isCreateAvgFace());
-        fpScalingCheckBox.setSelected(c.isFpScaling());
         fpThresholdSlider.setValue(c.getFpTreshold());
         exportLandmarksButton.setVisible(reg == RegistrationMethod.PROCRUSTES);
         procrustesPanel.setVisible(c.getComparisonMethod() == ComparisonMethod.PROCRUSTES);
@@ -583,7 +558,6 @@ public class OneToManyComparisonConfiguration extends javax.swing.JPanel {
     private javax.swing.JLabel createAvgLabel;
     private javax.swing.JButton exportLandmarksButton;
     private javax.swing.Box.Filler filler1;
-    private javax.swing.JCheckBox fpScalingCheckBox;
     private javax.swing.JSlider fpThresholdSlider;
     private javax.swing.JButton jButton1;
     private javax.swing.JButton jButton10;
@@ -591,7 +565,6 @@ public class OneToManyComparisonConfiguration extends javax.swing.JPanel {
     private javax.swing.JLabel jLabel12;
     private javax.swing.JLabel jLabel2;
     private javax.swing.JLabel jLabel3;
-    private javax.swing.JLabel jLabel8;
     private javax.swing.JPanel jPanel3;
     private javax.swing.JPanel jPanel5;
     private javax.swing.JScrollPane jScrollPane1;
