+++ GUI/src/cz/fidentis/gui/comparison_two_faces/Bundle.properties
@@ -41,14 +41,6 @@ PairComparisonResults.jLabel2.text=Comparison results
 PairComparisonResults.jButton3.text=Export visual results...
 PairComparisonResults.jButton7.text=Back
 PairComparisonConfiguration.jLabel2.text=Comparison
-PairComparisonConfiguration.jLabel12.text=Threshold:
-PairComparisonConfiguration.jButton1.text=Browse...
-PairComparisonConfiguration.jTextField1.text=
-PairComparisonConfiguration.createDatabaseRadioButton.text=create database from files
-PairComparisonConfiguration.chooseDatabaseRadioButton.text=choose database file
-PairComparisonConfiguration.deafultDatabaseRadioButton.text=use default database
-PairComparisonConfiguration.jLabel8.text=Enable scaling:
-PairComparisonConfiguration.jLabel7.text=Use database:
 PairComparisonConfiguration.jLabel3.text=Comparison:
 PairComparisonConfiguration.processComparisonButton.text=Process Comparison
 PairRegistrationConfiguration.jLabel10.text=Use scale:
@@ -121,8 +113,6 @@ PairRegistrationConfiguration.earsCheckbox.text=Ears
 PairRegistrationConfiguration.icpScaleCheckbox.text=
 PairRegistrationConfiguration.continueComparisonCheckbox.text=Continue directly with comparison
 PairComparisonConfiguration.exportLandmarksButton.text=Export registered landmarks...
-PairComparisonConfiguration.enableScaleCheckbox.text=
-PairComparisonConfiguration.useDatabaseCheckbox.text=
 PairComparisonConfiguration.primarySolidCheckbox.text=solid
 PairComparisonConfiguration.secondarySolidCheckbox.text=solid
 PairComparisonConfiguration.innerSurfaceSolidCheckbox.text=Inner surface solid
@@ -142,3 +132,11 @@ PairComparisonResults.innerSurfaceRadioButton.text=inner surface modification
 PairComparisonConfiguration.exportFacesButton.text=Export registered faces...
 PairComparisonConfiguration.backRegistrationButton.text=<- Edit registration criteria
 PairRegistrationConfiguration.registerButton.text=Register
+PairComparisonConfiguration.useDatabaseCheckbox.text=
+PairComparisonConfiguration.jLabel7.text=Use database:
+PairComparisonConfiguration.jLabel12.text=Threshold:
+PairComparisonConfiguration.jButton1.text=Browse...
+PairComparisonConfiguration.jTextField1.text=
+PairComparisonConfiguration.createDatabaseRadioButton.text=create database from files
+PairComparisonConfiguration.chooseDatabaseRadioButton.text=choose database file
+PairComparisonConfiguration.deafultDatabaseRadioButton.text=use default database
