+++ GUI/src/cz/fidentis/gui/comparison_two_faces/PairRegistrationConfiguration.form
@@ -1038,7 +1038,7 @@
                     </Property>
                   </Properties>
                   <Events>
-                    <EventHandler event="stateChanged" listener="javax.swing.event.ChangeListener" parameters="javax.swing.event.ChangeEvent" handler="fpScaleCheckBoxStateChanged"/>
+                    <EventHandler event="actionPerformed" listener="java.awt.event.ActionListener" parameters="java.awt.event.ActionEvent" handler="fpScaleCheckBoxActionPerformed"/>
                   </Events>
                 </Component>
               </SubComponents>
