+++ nbproject/genfiles.properties
@@ -8,4 +8,4 @@ nbproject/build-impl.xml.script.CRC32=bfb4b8ee
 nbproject/build-impl.xml.stylesheet.CRC32=0f381476@2.50.1
 nbproject/platform.xml.data.CRC32=4b80fb95
 nbproject/platform.xml.script.CRC32=6dcbd131
-nbproject/platform.xml.stylesheet.CRC32=4e1f53d4@2.62.1
+nbproject/platform.xml.stylesheet.CRC32=45ddf0e0@2.72.1
